+++ Changes.md
@@ -1,3 +1,4 @@
+* #726 - improve readability of subjective score summary
 * #711 - handle double quotes in team name on the select team page for score entry
 * #600 - migrate from FindBugs to SpotBugs
 * #724 - update GatherBugReport to find logs in the new location after switching to embedded Tomcat